refactor(cockpit): extract gate evidence module
